requirements.txt
langgraph
langchain
langchain-openai
langchain-qdrant
qdrant-client
python-dotenv
langchain-community
fastapi
uvicorn
pypdf
python-docx
python-magic-bin
unstructured
langsmith
openai
langgraph-cli[inmem]
requests